Fixes composite identifier not visible to JPA
I am not sure if this fix is correct.

The hibernate documentation [1] says that there are the following ways to define a composite identity:

- @EmbeddedId
- @IdClass
- Composite identifiers with associations => Use @Id multiple times

I am not sure what the binding by `bindCompositeId` corresponds to.

I just know that in the 3rd case hibernate does set the `identifierMapper`. I am not sure if there is valid in the other cases.

[1]: https://docs.jboss.org/hibernate/orm/5.2/userguide/html_single/Hibernate_User_Guide.html#identifiers-composite